% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_composition.R
\docType{data}
\name{agg_composition}
\alias{agg_composition}
\title{Composition of country aggregates included in the 2019 version of the EU's Handbook on the external costs of transport}
\format{
A tibble
\describe{
\item{iso3c}{Country code (ISO codes)}
\item{code}{Aggregate code}
\item{name}{Aggregate name}
\item{type}{Aggregate type}
}
}
\usage{
agg_composition
}
\description{
List of the countries making up the maritime aggregates for which the Handbook provides
pollutant cost factors.
}
\keyword{datasets}
