% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_transfer.R
\name{bt_transfer}
\alias{bt_transfer}
\title{Computing transfer factos for unit value trasnfer with income adjustment}
\usage{
bt_transfer(
  study_site = "EUU",
  policy_site,
  study_yr = 2016,
  policy_yr = 2019,
  aggregate = "no",
  currency = "EUR"
)
}
\arguments{
\item{study_site}{A single string. The name of the country in which the original
study that identified the values to be transferred was carried out. The name can be
provided in any language (first letter always capitalized). The default value is \code{EUU}
which statnds for the European Union.}

\item{policy_site}{A string vector. Names of the countries to which the value must be
transferred. It can be either a single string or a vector. Names can be provided in
any language.}

\item{study_yr}{Numeric. The year in which the original values was estimated. Default
is 2016, year about which the estimates provided by the European Commission's Handbook
on the external cost of the transport sector refer to.}

\item{policy_yr}{Numeric. The ywar to which values must be transferred. Default is 2019.}

\item{aggregate}{A single string. It can assume three different values. If \code{no} transfer
factors are provided for each country selected via \code{policy_site}. If \code{yes} a transfer
factor is calculated for the aggregate made up by the countries selected via \code{policy_site}.
If \code{row} a transfer factor is calculated for the aggregate "rest of the world" (world
minus countries selected via \code{policy_site}).}

\item{currency}{A single string. It can assume two values: \code{EUR} (the default) or
\code{USD}. It refers to the currency in which the original values is expressed.}
}
\value{
A data frame containing all parameters used and the transfer factor (bt_fct), which
is the scalar which the original value must be multiplied by to perform the transfer. Such
factor is already adjusted for inflation.
}
\description{
\code{bt_transfer} returns a data frame with all parameters used to calculate
transfer factors which can be in turn used to perform value transfers in
environmental and health economics.
}
\details{
This function allows the calculation of transfer factors to be used to perform
unit value transfers with income adjustments. Such factors are scalars that must be
multiplied by the original value. The function allows the user to select the
study site (country in which the original stusy that identified the value to be
transferred was carried out) and the policy sites. The latter can be a group of
countries or an aggregate of countries. Study and policy years can also be selected,
as well as whether the original estimate is expressed in US dollars or euro (default).
Finally, the function considers different values for income elasticity of the
willingness to pay for the environmental/health good to be transferred (epsilon)
according to the policy country's GNI per capita. More information on value
transfer techniques can be found in Navrud and Ready (2007).
}
\examples{
bt_transfer(policy_site = c("Italia", "Allemagne", "France", "Espa単a", "Polska"))
bt_transfer(study_site = "United States",
policy_site = c("Italia", "Allemagne", "France", "Espa単a", "Polska"),
currency = "USD")
bt_transfer(policy_site = c("Italia", "Allemagne", "France", "Espa単a", "Polska"), aggregate = "yes")
bt_transfer(policy_site = c("Italia", "Allemagne", "France", "Espa単a", "Polska"), aggregate = "row")
}
