% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btransfer.R
\name{btransfer}
\alias{btransfer}
\title{Computing transfer factors for unit value transfer with income adjustment}
\usage{
btransfer(
  study_site = "European Union",
  policy_site,
  study_yr = 2016,
  policy_yr,
  agg_policy = "no",
  study_currency = "EMU",
  policy_currency,
  avg_h = 20
)
}
\arguments{
\item{study_site}{A single string. The name of the country in which the original
study that identified the values to be transferred was carried out. The name can be
provided in any language (first letter always capitalized). It is possible to choose
also some aggregates: the European Union, the world as a whole and a few maritime aggregates present in
the 2019 version of the EU's Handbook on the External Costs of Transport:
- Mediterranean sea (`med`)
- Black sea (`blk`)
- North sea (`nor`)
- Atlantic ocean (`atl`), European countries only
- Baltic sea (`bal`)
Single country names can
be provided in English, French, Italian, Spanish, German, Dutch and Portuguese.
The default value is `European Union`.}

\item{policy_site}{A string vector. Names of the country to which the value must be
transferred. It is possible to choose amongst some aggregates: the Eu, the world and the maritime aggregates
(see `study_site`. If a vector of names is provided and the `agg_policy` argument
is set to `yes` or `row`, such vector is used to compute a value transfer factor to the
aggregate made up by the countries specified or the Rest of the world (world total minus the
countries specified). If one of the maritime aggregates is chosen, `agg_policy` must be set to `yes`.
Names can be provided in English, French, Spanish, Italian, German, Dutch and Portuguese.}

\item{study_yr}{Numeric. The year in which the original values was estimated. Default
is 2016, year about which the estimates provided by the European Commission's Handbook
on the external cost of the transport sector refer to.}

\item{policy_yr}{Numeric. The year to which values must be transferred. Default is 2019. It
is possible to define also future years up to 2050.}

\item{agg_policy}{A character string. It can assume three different values. If `no` transfer
factors are provided a single country selected via `policy_site`. If `yes` a transfer
factor is calculated for the aggregate made up by the countries selected via `policy_site`.
If `row` a transfer factor is calculated for the aggregate "rest of the world" (world
minus countries selected via `policy_site`). In case one of the maritime aggregates is chosen as
`policy_site`, this parameter must be set to `yes` to have the aggregate value, otherwise an error is returned.}

\item{study_currency}{A character string. It can be specified using the ISO code of the country
of interest. For example, US dollar is specified using `USA`, while the GB Pound is specified
with `GBR`. The euro is denoted with `EMU`. The GDP deflator used to take inflation into account
(when `study_yr` and `policy_yr` are not the same) depends on the `study_currency`. If
`study_currency = "USA` then the US GDP deflator is used, for example. If the ISO code is not known it
can be retrieved using the [iso_codes] function.}

\item{policy_currency}{A single string. A character string. It can be specified using the ISO code of the country
of interest. See `study_currency`.}

\item{avg_h}{A numeric value. Number of years to be considered to compute average growth rates for
GDP and GNI per capita in case `policy_yr` is in the future.}
}
\value{
A numeric value (scalar) representing the transfer factor (bt_fct), which
    is the number to which the original value must be multiplied by to perform the transfer. Such
    factor is already adjusted for inflation and expressed in the `policy_yr` price levels. If `policy_yr`
    is a year into the future, price levels corresponds to the last year for which data are available.
}
\description{
`btransfer` returns value factors to perform value transfers from a study to a policy site, in
environmental and health economics.
}
\details{
This function allows the calculation of transfer factors to be used to perform
unit value transfers with income adjustment. Such factors are scalars that must be
multiplied by the original value. The function allows the user to select the
study site (country in which the original study that identified the value to be
transferred was carried out) and the policy sites. Study and policy years can also be selected,
as well as whether the original estimate is expressed in US dollars or euro (default).
Finally, the function considers different values for income elasticity of the
willingness to pay for the environmental/health good to be transferred (epsilon)
according to the policy country's GNI per capita. More information on value
transfer techniques can be found in Navrud and Ready (2007).
The function works also in case policy year is set to a future year. Parameters like
GDP per capita and GNI per capita (Needed to identify the appropriate epsilon)
are calculated, for the policy year in the future, using average growth rates
of the last n-year period, which can be defined by the user.
}
\examples{
btransfer(policy_site = "Brazil", policy_yr = 2020, policy_currency = "BRA")
btransfer(study_site = "blk", policy_site = "med", policy_yr = 2020,
agg_policy = "yes", policy_currency = "EMU")
btransfer(study_site = "Italia", policy_site = c("Francia", "Germany", "Portugal"),
policy_yr = 2020, agg_policy = "row", policy_currency = "USA")
}
