% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btransfer_nuts2.R
\name{btransfer_nuts2}
\alias{btransfer_nuts2}
\title{Computing transfer factors for unit value transfer with income adjustment at NUTS2 level}
\usage{
btransfer_nuts2(
  study_site = "Italia",
  policy_site,
  study_yr = 2016,
  policy_yr,
  ref_yr,
  epsilon = 0.2,
  simplify = TRUE
)
}
\arguments{
\item{study_site}{A string. The name (in any language) of the EU country from which the value has to be
transferred. the European Union as a whole can spe specified in English and Italian only. Only
one `study_site` can be specified at once.}

\item{policy_site}{A character vector. A vector of Italian NUTS2 names to perform the transfer to. Names must
be in Italian. See the full list checking the documentation of `nuts2_codes`.}

\item{study_yr}{A numeric value. The study year.}

\item{policy_yr}{A numeric value. The policy year.}

\item{ref_yr}{A numeric value. The year at which prices must be expressed.}

\item{epsilon}{A numeric value. The income elasticity of the willingness to pay for
environmental/health goods. Default value is `0.2`.}

\item{simplify}{A logical value. If set to `TRUE` the function returns the transfer factor only. If set to
`FALSE` it returns a tibble with information on the selected NUTS.}
}
\value{
Either a single value (transfer factor) or A tibble containing transfer factors for the selected NUTS2 regions.
}
\description{
This function is a simplified version of `btransfer` which computes transfer factors
between any European country belonging to the EU (or the Eu itself) and any of the Italian regions.
In this simplified version it is not possible to transfer values to the future.
Epsilon, that is the income elasticity of the willingness to pay for
environmental/health goods, can be defined by the user.
}
\examples{

btransfer_nuts2(policy_site = "Umbria", policy_yr = 2019, ref_yr = 2019)
}
