% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_avg.R
\name{compute_avg}
\alias{compute_avg}
\title{Compute average values of macro-economic variables}
\usage{
compute_avg(iso, start_yr, end_yr, var, type = "levels")
}
\arguments{
\item{iso}{A string. A country's ISO code (3 digits). The function accepts
\code{WLD} and \code{EUU} for the world and the European Union respectively.}

\item{start_yr}{A double. Beginning of the time window over which the average
is computed.}

\item{end_yr}{A double. End of the time window over which the average
is computed.}

\item{var}{A string. The short name of the macro-economic variable of interest.
Possible values are \code{gdp}, \code{gdp_capita}, \code{gni},
\code{pop} and \code{death_rate}, \code{gni_capita}.}

\item{type}{A string. Indicates the type of data over which compute the average
values. It can assume two values: \code{levels} or \code{growth_rt}.}
}
\value{
A double. The average growth rate of the selected variable for the
country and time window of interest.
}
\description{
This function computes average values of some macro-economic variables
for selected countries and time intervals. Underlying data can be either
levels (\code{\link{wb_series}}) or growth rates (\code{\link{wb_growth}}).
}
