% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epsilon.R
\name{compute_epsilon}
\alias{compute_epsilon}
\title{Computes the income elasticity of the willingness to pay}
\usage{
compute_epsilon(gni_capita)
}
\arguments{
\item{gni_capita}{A double. The gross national income per capita of the
country for which epsilon must be estimated. Expressed in constant USD, at the
prices of the year in which the WB performed the income classification. See
\url{https://blogs.worldbank.org/opendata/new-world-bank-country-classifications-income-level-2021-2022}}
}
\value{
A double. The income elasticity of the willingness to pay
}
\description{
This function computes the income elasticity of the willingness to pay for
environmental/health goods (epsilon) based on GNI per capita and the latest
version of the World Bank's income classification of countries.
}
\examples{
compute_epsilon(30000)
}
\seealso{
\code{\link{income_class}}
}
