% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eta.R
\name{compute_eta}
\alias{compute_eta}
\title{Computing eta for social discount rate estimation}
\usage{
compute_eta(iso_code, policy_yr, h, eta_lit = 1.35)
}
\arguments{
\item{iso_code}{A string. An ISO code (3 digits) representing the country for which
eta is to be calculated. \code{OAVG} returns OECD's average values.}

\item{policy_yr}{A double. The last year (4 digits) used to compute the average mean
and marginal tax rates, over an h-long period of time.}

\item{h}{A double. Number of the years oever which computing average values.}

\item{eta_lit}{A double. A custom value for \code{eta}, in case a literature value needs
to be used, tipically for non-OECD countries.}
}
\value{
A double. The estimate of the elasticity of marginal utility of consumption
for the selected country, over an h-long period starting backwards from \code{policy_yr}
}
\description{
This function computes \code{eta}, the elasticity of marginal utility of consumption,
following the equal-sacrifice income tax approach to be used to estimate a
country's social discount rate using the Ramsey's rule.
Eta is estimated from marginal and average tax rates from the OECD tax database.
}
\examples{
compute_eta("ITA", 2021, 20)
}
