% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gdp_dfl.R
\name{compute_gdp_dfl}
\alias{compute_gdp_dfl}
\title{Computes the GDP deflator for a given country/currency and year}
\source{
\url{https://ec.europa.eu/eurostat/data/database} and \url{https://data.worldbank.org/}
}
\usage{
compute_gdp_dfl(country_iso = "EMU", ref_year)
}
\arguments{
\item{country_iso}{A character string. The ISO code (3 digits) of the country of interest.
Default is the Euro Area (\code{EMU}).}

\item{year}{A double. The year for which the GDP deflator must be calculated.
calculated.}
}
\value{
A numeric value: the GDP deflator for the country/year combination
selected.
}
\description{
This function computes the GDP deflator for a country provided by the user.
Data are from the World Bank database and cover most countries in the world.
Data on the Euro zone are from Eurostat.
}
\examples{
compute_gdp_dfl("ITA", 2019)
compute_gdp_dfl(year = 2014)
}
