% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_growth_rt.R
\name{compute_growth_rate}
\alias{compute_growth_rate}
\title{Compute growth rates of a vector of values}
\usage{
compute_growth_rate(x)
}
\arguments{
\item{x}{A vector of numeric values. It can handle \code{NA_real_}.}
}
\value{
A vector of growth rates
}
\description{
This function computes growth rates along a vector of values.
}
\examples{
x <- c(rep(NA_real_, 5), 200, 210, 215, 222, 230, 300)
compute_growth_rate(x)
}
