% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_macro_var.R
\name{compute_macro_var}
\alias{compute_macro_var}
\title{Returns macro-economic variables for selected countries and years}
\usage{
compute_macro_var(country_iso, ref_year, var = "gdp_capita", agg = "no")
}
\arguments{
\item{country_iso}{A string. The ISO code (3 digits) of the country of interest.
Codes \code{WLD} and \code{EUU} stand for the world and the European Union
respectively. In case \code{agg} is set to either \code{yes} or \code{row}
more than one iso code can be provided.}

\item{ref_year}{A double. Year of interest. Can be a year into the future up
to 2050. Years before 1961 are not accepted neither.}

\item{var}{A string. The name of the macro-economic variable to be extracted.
It can assume the following values: \code{gdp}, \code{gdp_capita}, \code{gni},
\code{pop}, \code{gni_capita} and \code{death_rate}.}

\item{agg}{A string. It can assume three values: \code{no}, the default, \code{yes}
and \code{row}. Indicates whether the variable of interest must be computed for an
aggregate. In case \code{agg = "row"}, the ROW is built substracting the selected
countries from the world total}
}
\value{
A double with the value of the selected variable for the country and year
of interest.
}
\description{
This function extracts and returns macro-economic variables from the
\code{wb_series} database, which are in turn retrieved from the
World Bank database. The user can specify the country and the year of interest.
In case the selected year is in the future, the function estimates future
values of the selected variable based on the average growth rates of the last
\code{n} years. Where \code{n} is the number of years between the selected
reference year and the latest year with available data. In case such difference
is lower than 5 years, 5 years is used a window to compute the average.
}
\seealso{
For more information on variables see the \code{wb_series} documentation.
}
