% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_list.R
\docType{data}
\name{country_list}
\alias{country_list}
\title{List of all individual countries iso3c codes}
\format{
An object of class \code{character} of length 218.
}
\source{
\url{https://data.worldbank.org/}
}
\usage{
country_list
}
\description{
A string vector made up by 3-digit iso3c codes of all countries in the world.
}
\seealso{
\code{\link[wbstats]{wb_countries}}
}
\keyword{datasets}
