% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_iso_to_name.R
\name{from_iso_to_name}
\alias{from_iso_to_name}
\title{Returns country names (in English) from ISO}
\usage{
from_iso_to_name(iso)
}
\arguments{
\item{iso}{A string. The a country's Iso code (3 digits). Special values
admitted are \code{WLD} (returns \code{World}) and \code{EUU} (returns
\code{European Union}). Function is based on \code{\link[countrycode]{countrycode}}.}
}
\value{
A string. The English name of the country of interest.
}
\description{
This function returns the English name of a country starting from its ISO code.
}
\seealso{
\code{\link[countrycode]{countrycode}}
}
