% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_defl.R
\name{gdp_defl}
\alias{gdp_defl}
\title{Computes the GDP deflator for a given country/currency and year}
\source{
\url{https://db.nomics.world/}
}
\usage{
gdp_defl(country_iso = "EMU", base_yr, ref_yr)
}
\arguments{
\item{country_iso}{A character string. The ISO code (3 digits) of the country of interest.
Default is the Euro Area `EMU`.}

\item{base_yr}{A numeric value. The year from which price levels the monetary value must be converted.}

\item{ref_yr}{A numeric value. The year to which price levels the monetary value must be converted.}
}
\value{
A numeric value: the factor which the monetary value to be converted must be multiplied by.
}
\description{
This function that returns a factor to convert monetary value from a base to a
reference year usding the GDP deflator of a given country/currency .
Data are from the World Bank and Eurostat databases and cover most countries in the world.
}
