% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_class.R
\docType{data}
\name{income_class}
\alias{income_class}
\title{World Bank country classification based in GNI per capita (Atlas method)}
\format{
A data frame
\describe{
\item{income_level}{Country income classification according to the WB}
\item{min}{Lower bound of income classification (international USD)}
\item{max}{Upper bound of income classification (international USD)}
\item{epsilon}{income elasticity of the willingness to pay of the corresponding income group}
}
}
\source{
\url{https://blogs.worldbank.org/opendata/new-world-bank-country-classifications-income-level-2021-2022}
}
\usage{
income_class
}
\description{
This dataset contains the latest available World Bank's income classification based
on gross national income (GNI) per capita (Atlas method). This classification is useful
to identify the most appropriate income elasticity of the willingness to pay for
the environmental/health good to be transferred (epsilon). Values are defined
on the basis of Navrud (2009) and email exchanges with Prof Navrud himself.
}
\keyword{datasets}
