% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_lvs.R
\docType{data}
\name{income_lvs}
\alias{income_lvs}
\title{World Bank country classification according to income levels}
\format{
a tibble
\describe{
\item{iso3c}{Three-digit country code}
\item{income_level_iso3c}{Income classification code}
\item{income_level}{Income classification - full name}
}
}
\source{
\url{https://data.worldbank.org/}
}
\usage{
income_lvs
}
\description{
The dataset contains the latest available country classification according to
income levels. The classification is based on GNI per capita and is updated
annually.
}
\keyword{datasets}
