% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes.R
\name{iso_codes}
\alias{iso_codes}
\title{Identifies ISO codes of a given country or aggregate}
\usage{
iso_codes(country_name)
}
\arguments{
\item{country_name}{A single string. The name of the country for which the
ISO code must be identified. Country names can be in any language. The function
returns ISO codes for the European Union ad the world as well, but, in these
cases, only English, French, Italian, Spanish, Portuguese, German and Dutch
are supported.}
}
\value{
A string, the ISO code of the country of interest.
}
\description{
This function exploits the \code{countrycode} package to identify the ISO
codes (3 digits) of a given country.
}
\seealso{
\code{\link[countrycode]{countryname}}
}
