% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts2_codes.R
\docType{data}
\name{nuts2_codes}
\alias{nuts2_codes}
\title{Eurostat codes of Italian NUTS2 regions}
\format{
A tibble
\describe{
\item{label}{Region name}
\item{code}{Region code}
}
}
\usage{
nuts2_codes
}
\description{
Eurostat codes of Italian NUTS2 regions
}
\keyword{datasets}
