% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_data.R
\docType{data}
\name{tax_data}
\alias{tax_data}
\title{Average and marginal income tax rates of OECD countries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1760 rows and 4 columns.
}
\source{
\url{https://www.oecd.org/ctp/tax-policy/tax-database/}
}
\usage{
tax_data
}
\description{
This dataset contains data on average and marginal income tax rates of OECD
countries retrieved from the OECD tax database.Data refer to a single person that
earns the 100% of the average income at national level.
}
\details{
\describe{
\item{iso3c}{Country codes (ISO3C), string.}
\item{year}{Year, numeric.}
\item{indicator}{2_5 stands for average tax rate, 3_1 for marginal tax rate.}
\item{value}{indicator value (in percentage).}
}
}
\keyword{datasets}
