% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_growth.R, R/wb_series.R
\docType{data}
\name{wb_series}
\alias{wb_series}
\title{World Bank data to be used in unit value transfer - growth rates}
\format{
A tibble
\describe{
\item{iso3c}{Country code (Iso 3 digits)}
\item{country}{Country name in English}
\item{year}{year}
\item{gdp_growth}{Annual real GDP growth (WB code: NY.GNP.MKTP.KD.ZG)}
\item{gdp_capita_growth}{Annual real GDP per capita growth (WB code: NY.GDP.PCAP.KD.ZG)}
\item{gni_growth}{Annual real GNI growth (WB code: NY.GNP.MKTP.KD.ZG)}
\item{pop_growth}{Annual population growth rate (Wb code: SP.POP.GROW)}
}

A tibble
\describe{
\item{iso3c}{Country code (Iso 3 digits)}
\item{country}{Country name in English}
\item{year}{year}
\item{gdp_defl}{GDP deflator (WB code: NY.GDP.DEFL.ZS)}
\item{gdp}{GDP at constant prices PPP (WB code: NY.GDP.MKTP.PP.KD)}
\item{gdp_capita}{GDP per capita at constant prices PPP (WB code: NY.GDP.PCAP.PP.KD)}
\item{gni}{GNI at current prices, Atlas method (Wb code: NY.GNP.ATLS.CD)}
\item{exc_rate}{Exchange rate, USD per LCU (WB code: PA.NUS.FCRF)}
\item{pop}{Total population (WB code: SP.POP.TOTL)}
}
}
\source{
\url{https://data.worldbank.org/}

\url{https://data.worldbank.org/}
}
\usage{
wb_series

wb_series
}
\description{
This dataset contains annual growth rates for the indicators used in unit
value transfer. Such data are used to "forecast" future indicator lavels to
perform benefit transfer to years in the future.

This dataset contains the latest available WB data on GDP and GDP per capita
(constant international dollars), GNI (current prices), population, GDP deflators and
exchange rates (USD per unit of LCU). Such data are used to perform unit value
transfer with income adjustment.
}
\keyword{datasets}
