% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_growth_rate.R
\name{compute_growth_rate}
\alias{compute_growth_rate}
\title{Compute growth rates of a vector of values}
\usage{
compute_growth_rate(x, avg = TRUE)
}
\arguments{
\item{x}{A vector of numeric values. It can handle \code{NA_real_} values.}

\item{avg}{A logical value. If `TRUE` (the default), the average growth rate is
compute. If `FALSE`, the whole vector of growth rates is returned.}
}
\value{
A vector of growth rates or their mean value
}
\description{
This function computes growth rates along a vector of values.
}
