% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_macro_var.R
\name{compute_macro_var}
\alias{compute_macro_var}
\title{Returns macro-economic variables for selected countries and years}
\usage{
compute_macro_var(
  country_iso,
  ref_yr,
  var = "gdp_capita",
  agg = "no",
  growth_rate_int = 20,
  avg = FALSE,
  growth_rt = FALSE
)
}
\arguments{
\item{country_iso}{A string. The ISO code (3 digits) of the country of interest.
Codes `WLD` and `EUU` stand for the world and the European Union
respectively. In case `agg` is set to either `yes` or `row`
more than one iso code can be provided.}

\item{ref_yr}{A double. Year of interest. Can be a year into the future up
to 2050. Years before 1961 are not accepted neither.}

\item{var}{A string. The name of the macro-economic variable to be extracted.
It can assume the following values: `gdp`, `gdp_capita`, `gni`,
`pop`, `gni_capita` and `death_rate`. GDP and GDP per capita are returned in
constant international dollars (2017). GNI and GNI per capita are returned in
constant USD. The base year is the same of the latest version of the WB income
classification of countries, since GNI per capita is essentially used to compute
the income elasticity of the marginal utility of consumption (`epsilon`).}

\item{agg}{A string. It can assume three values: `no`, the default, `yes`
and `row`. Indicates whether the variable of interest must be computed for an
aggregate. In case `row`, the ROW is built as the difference between the selected
countries from the world total.}

\item{growth_rate_int}{a numeric value. The number of years to be considered to calculate
average growth rates for the variable of interest in case `ref_yr` is set into the
future.}

\item{avg}{A logical value. If `TRUE` the average value of the variable specified with
`var` in the time window specified via `growth_rate_int` is returned. Default is set
to `FALSE`. If `ref_yr` is in the future, the average value returned is always computed
starting from historical data, up to the last available year.}

\item{growth_rt}{A logical value. If `TRUE` the average growth rate of the variable
specified with `var` is computed over the time interval specified through `growth_rate_int`.
If `ref_yr` is in the future, the average growth rate returned is always computed
starting from historical data, up to the last available year.}
}
\value{
A list with the value of the selected variable for the country and year
    of interest (`value`) and the latest year of available data used in the computation (`last_value`)
}
\description{
This function extracts and returns macro-economic variables from the
World Bank database (via DbNomics). The user can specify the country and the year of interest.
In case the selected year is in the future, the function estimates future
values of the selected variable based on the average growth rates of the last
`n` years (a parameter that can, in turn, be chosen by the user).
The function downloads the most recent data avilable.
}
