% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_macro_var.R
\name{compute_macro_var}
\alias{compute_macro_var}
\title{Returns macro-economic variables for selected countries and years}
\usage{
compute_macro_var(country_iso, ref_year, var = "gdp_capita")
}
\arguments{
\item{country_iso}{A string. The ISO code (3 digits) of the country of interest.
Codes \code{WLD} and \code{EUU} stand for the world and the European Union
respectively.}

\item{ref_year}{A double. Year of interest. Can be a year into the future up
to 2050. Years before 1961 are not accepted neither.}

\item{var}{A string. The name of the macro-economic variable to be extracted.
It can assume the following values: \code{gdp}, \code{gdp_capita}, \code{gni},
\code{pop} and \code{gni_capita}, \code{death_rate}.}
}
\value{
A double with the value of the selected variable for the country and year
of interest.
}
\description{
This function extracts and returns macro-economic variables from the
\code{\link{wb_series}} database, which are in turn retrieved from the
World Bank database. The user can specify the country and the year of interest.
In case the selected year is in the future, the function estimates future
values of the selected variable based on the average growth rates of the last
\code{n} years. Where \code{n} is the number of years between the selected
reference year and the latest year with available data. IN case such difference
is lower than 5 years, 5 years is used a window to compute the average.
}
\examples{
compute_macro_var("ITA", 2016, "gdp")
compute_macro_var("USA", 2030, "gdp")
}
\seealso{
\code{\link{wb_series}} for more information on variables.
}
